Allow disabling of xkb
The `xkbcommon` library is geared towards platforms using the XKB specification so it should be disableable. Add a WPE_DISABLE_XKB_INPUT option to prevent xkb code from being compiled. By default this option is off so this is not a breaking change.

The contents of `wpe/input.h` that require `xkbcommon` are split into `wpe/input-xkb.h` The `src/input.c` was renamed to `src/input-xkb.c` to indicate the file is specific to xkb being enabled.
